Consolidate docker and integration test helper functions
