+++ tests/docker/test-helpers.go
@@ -1,7 +1,6 @@
 package docker
 
 import (
-	"context"
 	"fmt"
 	"math/rand"
 	"net"
@@ -15,11 +14,6 @@ import (
 
 	"golang.org/x/mod/semver"
 	"golang.org/x/sync/errgroup"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/utils/set"
 )
 
 type TestConfig struct {
@@ -490,122 +484,4 @@ func (config TestConfig) DeployWorkload(workload string) (string, error) {
 		}
 	}
 	return "", nil
-}
-
-// TODO the below functions are duplicated in the integration test utils. Consider combining into commmon package
-
-// CheckDefaultDeployments checks if the default deployments: coredns, local-path-provisioner, metrics-server, traefik
-// for K3s are ready, otherwise returns an error
-func CheckDefaultDeployments(kubeconfigFile string) error {
-	return DeploymentsReady([]string{"coredns", "local-path-provisioner", "metrics-server", "traefik"}, kubeconfigFile)
-}
-
-// DeploymentsReady checks if the provided list of deployments are ready, otherwise returns an error
-func DeploymentsReady(deployments []string, kubeconfigFile string) error {
-
-	deploymentSet := make(map[string]bool)
-	for _, d := range deployments {
-		deploymentSet[d] = false
-	}
-
-	client, err := k8sClient(kubeconfigFile)
-	if err != nil {
-		return err
-	}
-	deploymentList, err := client.AppsV1().Deployments("").List(context.Background(), metav1.ListOptions{})
-	if err != nil {
-		return err
-	}
-	for _, deployment := range deploymentList.Items {
-		if _, ok := deploymentSet[deployment.Name]; ok && deployment.Status.ReadyReplicas == deployment.Status.Replicas {
-			deploymentSet[deployment.Name] = true
-		}
-	}
-	for d, found := range deploymentSet {
-		if !found {
-			return fmt.Errorf("failed to deploy %s", d)
-		}
-	}
-
-	return nil
-}
-
-func ParseNodes(kubeconfigFile string) ([]corev1.Node, error) {
-	clientSet, err := k8sClient(kubeconfigFile)
-	if err != nil {
-		return nil, err
-	}
-	nodes, err := clientSet.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return nodes.Items, nil
-}
-
-func ParsePods(kubeconfigFile string) ([]corev1.Pod, error) {
-	clientSet, err := k8sClient(kubeconfigFile)
-	if err != nil {
-		return nil, err
-	}
-	pods, err := clientSet.CoreV1().Pods("").List(context.Background(), metav1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	return pods.Items, nil
-}
-
-// PodReady checks if a pod is ready by querying its status
-func PodReady(podName, namespace, kubeconfigFile string) (bool, error) {
-	clientSet, err := k8sClient(kubeconfigFile)
-	if err != nil {
-		return false, err
-	}
-	pod, err := clientSet.CoreV1().Pods(namespace).Get(context.TODO(), podName, metav1.GetOptions{})
-	if err != nil {
-		return false, fmt.Errorf("failed to get pod: %v", err)
-	}
-	// Check if the pod is running
-	for _, containerStatus := range pod.Status.ContainerStatuses {
-		if containerStatus.Name == podName && containerStatus.Ready {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-// Checks if provided nodes are ready, otherwise returns an error
-func NodesReady(kubeconfigFile string, nodeNames []string) error {
-	nodes, err := ParseNodes(kubeconfigFile)
-	if err != nil {
-		return err
-	}
-	nodesToCheck := set.New(nodeNames...)
-	readyNodes := make(set.Set[string], 0)
-	for _, node := range nodes {
-		for _, condition := range node.Status.Conditions {
-			if condition.Type == corev1.NodeReady && condition.Status != corev1.ConditionTrue {
-				return fmt.Errorf("node %s is not ready", node.Name)
-			}
-			readyNodes.Insert(node.Name)
-		}
-	}
-	// Check if all nodes are ready
-	if !nodesToCheck.Equal(readyNodes) {
-		return fmt.Errorf("expected nodes %v, found %v", nodesToCheck, readyNodes)
-	}
-	return nil
-}
-
-func k8sClient(kubeconfigFile string) (*kubernetes.Clientset, error) {
-	config, err := clientcmd.BuildConfigFromFlags("", kubeconfigFile)
-	if err != nil {
-		return nil, err
-	}
-	clientSet, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return clientSet, nil
 }
